+++ Moco/Bindings/Java/Matlab/Examples/example2DWalking/example2DWalking.m
@@ -86,9 +86,9 @@ track.set_track_reference_position_derivatives(true);
 track.set_apply_tracked_states_to_guess(true);
 track.set_initial_time(0.0);
 track.set_final_time(0.47008941);
+
 study = track.initialize();
 problem = study.updProblem();
-
 
 % Goals
 % =====