+++ Moco/Bindings/Python/examples/example3DWalking/exampleMocoInverse.py
@@ -43,6 +43,20 @@ def solveMocoInverse():
     # Only valid for DeGrooteFregly2016Muscles.
     modelProcessor.append(osim.ModOpScaleActiveFiberForceCurveWidthDGF(1.5))
     modelProcessor.append(osim.ModOpAddReserves(1.0))
+    modelProcessor.append(osim.ModOpAppliesForce(False, [
+        "/forceset/contactSphereHeel_r",
+        "/forceset/contactLateralRearfoot_r",
+        "/forceset/contactLateralMidfoot_r",
+        "/forceset/contactLateralToe_r",
+        "/forceset/contactMedialToe_r",
+        "/forceset/contactMedialMidfoot_r",
+        "/forceset/contactSphereHeel_l",
+        "/forceset/contactLateralRearfoot_l",
+        "/forceset/contactLateralMidfoot_l",
+        "/forceset/contactLateralToe_l",
+        "/forceset/contactMedialToe_l",
+        "/forceset/contactMedialMidfoot_l"
+    ]))
     inverse.setModel(modelProcessor)
 
     # Construct a TableProcessor of the coordinate data and pass it to the
@@ -84,6 +98,20 @@ def solveMocoInverseWithEMG():
     modelProcessor.append(osim.ModOpIgnorePassiveFiberForcesDGF())
     modelProcessor.append(osim.ModOpScaleActiveFiberForceCurveWidthDGF(1.5))
     modelProcessor.append(osim.ModOpAddReserves(1.0))
+    modelProcessor.append(osim.ModOpAppliesForce(False, [
+        "/forceset/contactSphereHeel_r",
+        "/forceset/contactLateralRearfoot_r",
+        "/forceset/contactLateralMidfoot_r",
+        "/forceset/contactLateralToe_r",
+        "/forceset/contactMedialToe_r",
+        "/forceset/contactMedialMidfoot_r",
+        "/forceset/contactSphereHeel_l",
+        "/forceset/contactLateralRearfoot_l",
+        "/forceset/contactLateralMidfoot_l",
+        "/forceset/contactLateralToe_l",
+        "/forceset/contactMedialToe_l",
+        "/forceset/contactMedialMidfoot_l"
+    ]))
     inverse.setModel(modelProcessor)
     inverse.setKinematics(osim.TableProcessor('coordinates.sto'))
     inverse.set_initial_time(0.81)