+++ Moco/Examples/C++/example3DWalking/exampleMocoTrack.cpp
@@ -16,10 +16,11 @@
  * limitations under the License.                                             *
  * -------------------------------------------------------------------------- */
 
+// TODO update comments for contact tracking
 /// This example features two different tracking problems solved using the
 /// MocoTrack tool. 
 ///  - The first problem demonstrates the basic usage of the tool interface
-///    to solve a torque-driven marker tracking problem. 
+///    to solve a torque-driven marker tracking problem.
 ///  - The second problem shows how to customize a muscle-driven state tracking 
 ///    problem using more advanced features of the tool interface.
 ///
@@ -30,7 +31,8 @@
 
 using namespace OpenSim;
 
-void torqueDrivenMarkerTracking() {
+void torqueDrivenMarkerTracking(
+        bool useFootGroundContact, bool enableContactTracking) {
 
     // Create and name an instance of the MocoTrack tool.
     MocoTrack track;
@@ -41,18 +43,56 @@ void torqueDrivenMarkerTracking() {
     // ModelOperators. Operations are performed in the order that they are
     // appended to the model. In C++, you may use the pipe operator '|' to 
     // append ModelOperators.
-    track.setModel(
+    auto modelProcessor =
             // Create the base Model by passing in the model file.
             ModelProcessor("subject_walk_armless.osim") |
-            // Add ground reaction external loads in lieu of a ground-contact
-            // model.
-            ModOpAddExternalLoads("grf_walk.xml") |
             // Remove all the muscles in the model's ForceSet.
             ModOpRemoveMuscles() |
             // Add CoordinateActuators to the model degrees-of-freedom. This
             // ignores the pelvis coordinates which already have residual 
             // CoordinateActuators.
-            ModOpAddReserves(250));
+            ModOpAddReserves(250);
+
+    std::vector<std::string> forceNamesRightFoot = {
+        "forceset/contactHeel_r",
+        "forceset/contactLateralRearfoot_r",
+        "forceset/contactLateralMidfoot_r",
+        "forceset/contactLateralToe_r",
+        "forceset/contactMedialToe_r",
+        "forceset/contactMedialMidfoot_r"};
+    std::vector<std::string> forceNamesLeftFoot = {
+        "forceset/contactHeel_l",
+        "forceset/contactLateralRearfoot_l",
+        "forceset/contactLateralMidfoot_l",
+        "forceset/contactLateralToe_l",
+        "forceset/contactMedialToe_l",
+        "forceset/contactMedialMidfoot_l"};
+    if (useFootGroundContact) {
+        if (enableContactTracking) {
+            // Configure the existing MocoContactTrackingGoal in MocoTrack.
+            auto& contactTracking = track.updContactTrackingGoal();
+            contactTracking.setEnabled(true);
+            contactTracking.setWeight(1e-4 / 2*forceNamesRightFoot.size());
+            contactTracking.setExternalLoadsFile("grf_walk.xml");
+            contactTracking.addContactGroup(forceNamesRightFoot, "Right_GRF");
+            contactTracking.addContactGroup(forceNamesLeftFoot, "Left_GRF");
+            // Track only the sagittal plane reaction forces.
+            contactTracking.setProjection("plane");
+            contactTracking.setProjectionVector(SimTK::Vec3(0, 0, 1));
+        }
+    } else {
+        // Add ground reaction external loads in lieu of a ground-contact
+        // model.
+        modelProcessor.append(ModOpAddExternalLoads("grf_walk.xml"));
+        // Disable the contact force elements in the model.
+        auto forceNamesAll = forceNamesRightFoot;
+        forceNamesAll.insert(forceNamesAll.end(), forceNamesLeftFoot.begin(),
+                forceNamesLeftFoot.end());
+        modelProcessor.append(ModOpAppliesForce(false, forceNamesAll));
+    }
+
+    track.setModel(modelProcessor);
+    modelProcessor.process().print("subject_walk_armless_torque_driven.osim");
 
     // Use this convenience function to set the MocoTrack markers reference
     // directly from a TRC file. By default, the markers data is filtered at
@@ -97,7 +137,8 @@ void torqueDrivenMarkerTracking() {
     MocoSolution solution = track.solve(true);
 }
 
-void muscleDrivenStateTracking() {
+void muscleDrivenStateTracking(
+        bool useFootGroundContact, bool enableContactTracking) {
 
     // Create and name an instance of the MocoTrack tool.
     MocoTrack track;
@@ -107,24 +148,74 @@ void muscleDrivenStateTracking() {
     // muscles in the model are replaced with optimization-friendly
     // DeGrooteFregly2016Muscles, and adjustments are made to the default muscle
     // parameters.
-    ModelProcessor modelProcessor =
+    //ModelProcessor modelProcessor =
+    //        ModelProcessor("subject_walk_armless.osim") |
+    //        ModOpAddExternalLoads("grf_walk.xml") |
+    //        ModOpIgnoreTendonCompliance() |
+    //        ModOpReplaceMusclesWithDeGrooteFregly2016() |
+    //        // Only valid for DeGrooteFregly2016Muscles.
+    //        ModOpIgnorePassiveFiberForcesDGF() |
+    //        // Only valid for DeGrooteFregly2016Muscles.
+    //        ModOpScaleActiveFiberForceCurveWidthDGF(1.5);
+
+    auto modelProcessor =
+            // Create the base Model by passing in the model file.
             ModelProcessor("subject_walk_armless.osim") |
-            ModOpAddExternalLoads("grf_walk.xml") |
-            ModOpIgnoreTendonCompliance() |
-            ModOpReplaceMusclesWithDeGrooteFregly2016() |
-            // Only valid for DeGrooteFregly2016Muscles.
-            ModOpIgnorePassiveFiberForcesDGF() |
-            // Only valid for DeGrooteFregly2016Muscles.
-            ModOpScaleActiveFiberForceCurveWidthDGF(1.5);
+            // Remove all the muscles in the model's ForceSet.
+            ModOpRemoveMuscles();// |
+            // Add CoordinateActuators to the model degrees-of-freedom. This
+            // ignores the pelvis coordinates which already have residual
+            // CoordinateActuators.
+            //ModOpAddReserves(250);
+
+    std::vector<std::string> forceNamesRightFoot = {"forceset/contactHeel_r",
+            "forceset/contactLateralRearfoot_r",
+            "forceset/contactLateralMidfoot_r", "forceset/contactLateralToe_r",
+            "forceset/contactMedialToe_r", "forceset/contactMedialMidfoot_r"};
+    std::vector<std::string> forceNamesLeftFoot = {"forceset/contactHeel_l",
+            "forceset/contactLateralRearfoot_l",
+            "forceset/contactLateralMidfoot_l", "forceset/contactLateralToe_l",
+            "forceset/contactMedialToe_l", "forceset/contactMedialMidfoot_l"};
+    if (useFootGroundContact) {
+        if (enableContactTracking) {
+            // Configure the existing MocoContactTrackingGoal in MocoTrack.
+            auto& contactTracking = track.updContactTrackingGoal();
+            contactTracking.setEnabled(true);
+            contactTracking.setWeight(1e-4 / 2 * forceNamesRightFoot.size());
+            contactTracking.setExternalLoadsFile("grf_walk.xml");
+            contactTracking.addContactGroup(forceNamesRightFoot, "Right_GRF");
+            contactTracking.addContactGroup(forceNamesLeftFoot, "Left_GRF");
+            // Track only the sagittal plane reaction forces.
+            contactTracking.setProjection("plane");
+            contactTracking.setProjectionVector(SimTK::Vec3(0, 0, 1));
+        }
+    } else {
+        // Add ground reaction external loads in lieu of a ground-contact
+        // model.
+        modelProcessor.append(ModOpAddExternalLoads("grf_walk.xml"));
+        // Disable the contact force elements in the model.
+        auto forceNamesAll = forceNamesRightFoot;
+        forceNamesAll.insert(forceNamesAll.end(), forceNamesLeftFoot.begin(),
+                forceNamesLeftFoot.end());
+        modelProcessor.append(ModOpAppliesForce(false, forceNamesAll));
+    }
+
     track.setModel(modelProcessor);
+    modelProcessor.process().print("subject_walk_armless_torque_driven.osim");
 
     // Construct a TableProcessor of the coordinate data and pass it to the 
     // tracking tool. TableProcessors can be used in the same way as
     // ModelProcessors by appending TableOperators to modify the base table.
     // A TableProcessor with no operators, as we have here, simply returns the
     // base table.
-    track.setStatesReference(TableProcessor("coordinates.sto"));
+    auto tableProcessor = TableProcessor("coordinates.sto");
+    track.setStatesReference(tableProcessor);
     track.set_states_global_tracking_weight(10);
+
+    MocoWeightSet stateWeights;
+    stateWeights.cloneAndAppend({"/jointset/ground_pelvis/pelvis_ty/value", 0});
+    stateWeights.cloneAndAppend({"/jointset/ground_pelvis/pelvis_tz/value", 0});
+    track.set_states_weight_set(stateWeights);
 
     // This setting allows extra data columns contained in the states
     // reference that don't correspond to model coordinates.
@@ -138,7 +229,7 @@ void muscleDrivenStateTracking() {
     // Initial time, final time, and mesh interval.
     track.set_initial_time(0.81);
     track.set_final_time(1.65);
-    track.set_mesh_interval(0.08);
+    track.set_mesh_interval(0.05);
 
     // Instead of calling solve(), call initialize() to receive a pre-configured
     // MocoStudy object based on the settings above. Use this to customize the
@@ -148,20 +239,31 @@ void muscleDrivenStateTracking() {
     // Get a reference to the MocoControlGoal that is added to every MocoTrack
     // problem by default.
     MocoProblem& problem = study.updProblem();
-    MocoControlGoal& effort =
-        dynamic_cast<MocoControlGoal&>(problem.updGoal("control_effort"));
+
+
+    TimeSeriesTable refTable("muscle_driven_state_tracking_tracked_states.sto");
+    auto startIndex = refTable.getNearestRowIndexForTime(0.81);
+    for (auto refLabel : refTable.getColumnLabels()) {
+        if (refLabel.find("/jointset/ground_pelvis") != std::string::npos) {
+            auto refCol = refTable.getDependentColumn(refLabel);
+            problem.setStateInfo(refLabel, {}, refCol[startIndex]);
+        }
+    }
+
+    //MocoControlGoal& effort =
+    //    dynamic_cast<MocoControlGoal&>(problem.updGoal("control_effort"));
 
     // Put a large weight on the pelvis CoordinateActuators, which act as the
     // residual, or 'hand-of-god', forces which we would like to keep as small
     // as possible.
-     Model model = modelProcessor.process();
-     for (const auto& coordAct : model.getComponentList<CoordinateActuator>()) {
-        auto coordPath = coordAct.getAbsolutePathString();
-        if (coordPath.find("pelvis") != std::string::npos) {
-            effort.setWeightForControl(coordPath, 10);
-        }
-    }
-    
+    //Model model = modelProcessor.process();
+    //for (const auto& coordAct : model.getComponentList<CoordinateActuator>()) {
+    //   auto coordPath = coordAct.getAbsolutePathString();
+    //   if (coordPath.find("pelvis") != std::string::npos) {
+    //       effort.setWeightForControl(coordPath, 10);
+    //   }
+    //}
+    //
     // Solve and visualize.
     MocoSolution solution = study.solve();
     study.visualize(solution);
@@ -169,15 +271,18 @@ void muscleDrivenStateTracking() {
 
 int main() {
 
+    bool useFootGroundContact = true;
+    bool enableContactTracking = false;
+
     // Solve the torque-driven marker tracking problem.
     // This problem takes a few minutes to solve.
-    torqueDrivenMarkerTracking();
+    //torqueDrivenMarkerTracking(useFootGroundContact, enableContactTracking);
 
     // Solve the muscle-driven state tracking problem.
     // This problem could take an hour or more to solve, depending on the 
     // number of processor cores available for parallelization. With 12 cores,
     // it takes around 25 minutes.
-    muscleDrivenStateTracking();
+    muscleDrivenStateTracking(useFootGroundContact, enableContactTracking);
 
     return EXIT_SUCCESS;
 }